refact waiting method to wait.go
