+++ pkg/runtime/runtime.go
@@ -11,7 +11,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"io"
-	"net"
 	nethttp "net/http"
 	"os"
 	"reflect"
@@ -1898,84 +1897,6 @@ func (a *DaprRuntime) getSecretStore(storeName string) secretstores.SecretStore 
 		return nil
 	}
 	return a.secretStores[storeName]
-}
-
-func (a *DaprRuntime) blockUntilAppPortOpen() {
-	if a.runtimeConfig.ApplicationPort <= 0 {
-		return
-	}
-
-	log.Infof("application protocol: %s. waiting on port %v.  This will block until the app is listening on that port.", string(a.runtimeConfig.ApplicationProtocol), a.runtimeConfig.ApplicationPort)
-
-	for {
-		conn, _ := net.DialTimeout("tcp", net.JoinHostPort("localhost", fmt.Sprintf("%v", a.runtimeConfig.ApplicationPort)), time.Millisecond*500)
-		if conn != nil {
-			conn.Close()
-			break
-		}
-		// prevents overwhelming the OS with open connections
-		time.Sleep(time.Millisecond * 50)
-	}
-
-	log.Infof("application discovered on port %v", a.runtimeConfig.ApplicationPort)
-}
-
-func (a *DaprRuntime) blockUntilAppIsReady() {
-
-	if !a.runtimeConfig.EnableWaitAppReady {
-		return
-	}
-
-	if a.runtimeConfig.Mode == modes.KubernetesMode {
-		log.Infof("going to get containers status : namespace = %s, podName = %s", a.namespace, a.podName)
-	out:
-		for {
-			// check readiness from k8s env
-			resp, err := a.operatorClient.GetContainersStatus(context.Background(), &operatorv1pb.GetContainersStatusRequest{
-				Namespace: a.namespace,
-				Name:      a.podName,
-			})
-
-			if err != nil {
-				log.Errorf("get containers status error : namespace = %s, podName = %s, err = %s", a.namespace, a.podName, err)
-				return
-			}
-
-			for name, ready := range resp.Statuses {
-				if name != "daprd" && !ready {
-					log.Infof("container[%s] not ready", name)
-					time.Sleep(time.Millisecond * 1000)
-					continue out
-				}
-			}
-
-			log.Infof("all containers are ready")
-			break
-		}
-	}
-
-	readinessAddress := a.runtimeConfig.ReadinessAddress
-	if readinessAddress == "" {
-		return
-	}
-
-	for {
-		req := invokev1.NewInvokeMethodRequest(a.runtimeConfig.ReadinessAddress)
-		req.WithHTTPExtension(nethttp.MethodGet, "")
-		req.WithRawData(nil, invokev1.JSONContentType)
-
-		ctx := context.Background()
-		resp, err := a.appChannel.InvokeMethod(ctx, req)
-
-		if err == nil && resp.Status().Code == nethttp.StatusOK {
-			log.Infof("application discovered on readiness address %s", a.runtimeConfig.ReadinessAddress)
-			break
-		}
-
-		time.Sleep(time.Millisecond * 200)
-	}
-
-	log.Infof("wait application ready, continue init")
 }
 
 func (a *DaprRuntime) loadAppConfiguration() {
